+++ wtforms/widgets/core.py
@@ -13,15 +13,22 @@ __all__ = (
 
 def html_params(**kwargs):
     """
-    Generate HTML parameters from inputted keyword arguments.
+    Generate HTML attribute syntax from inputted keyword arguments.
 
     The output value is sorted by the passed keys, to provide consistent output
-    each time this function is called with the same parameters.  Because of the
+    each time this function is called with the same parameters. Because of the
     frequent use of the normally reserved keywords `class` and `for`, suffixing
     these with an underscore will allow them to be used.
 
-    >>> html_params(name='text1', id='f', class_='text') == 'class="text" id="f" name="text1"'
-    True
+    In addition, the values ``True`` and ``False`` are special:
+      * ``attr=True`` generates the HTML compact output of a boolean attribute,
+        e.g. ``checked=True`` will generate simply ``checked``
+      * ``attr=`False`` will be ignored and generate no output.
+
+    >>> html_params(name='text1', id='f', class_='text')
+    'class="text" id="f" name="text1"'
+    >>> html_params(checked=True, readonly=False, name="text1", abc="hello")
+    'abc="hello" checked name="text1"'
     """
     params = []
     for k, v in sorted(iteritems(kwargs)):