+++ CHANGES.rst
@@ -12,6 +12,7 @@ Not yet released.
 - Fix issue rendering SelectFields with ``value=True``
 - Make `DecimalField` able to use babel locale-based number formatting.
 - Drop Python 3.2 support (Python3 support for 3.3+ only)
+- passing ``attr=False`` to WTForms widgets causes the value to be ignored.
 
 Version 1.0.5
 -------------