+++ docs/whats_new.rst
@@ -13,6 +13,8 @@ New Features
 * :doc:`Class Meta <meta>` paradigm allows customization of many aspects of WTForms.
 * :doc:`CSRF <csrf>` and :doc:`i18n <i18n>` are core features not needing 
   extensions anymore.
+* Passing ``<attribute name>=False`` to WTForms widget rendering is now ignored,
+  making it easier to deal with boolean HTML attributes.
 
 
 Deprecated API's